+++ src/main/java/com/example/conflicttracker/repository/EventRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface EventRepository extends JpaRepository<Event, Integer> {
+public interface EventRepository extends JpaRepository<Event, Long> {
 
-    List<Event> findByConflictId(int conflictId);
+    List<Event> findByConflictId(Long conflictId);
 
 }
